% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cif.R
\name{auc.ws}
\alias{auc.ws}
\title{AUC for Weighted CIF}
\usage{
auc.ws(data, eta, tau)
}
\arguments{
\item{data}{A data frame with structure described in \code{\link{w_surv}}.}

\item{eta}{The event type of interest.}

\item{tau}{The cutoff on the weighted time scale (\code{wU}).}
}
\value{
Numeric value of the weighted AUC.
}
\description{
Computes the area under the weighted survival function curve for event type \code{eta}.
}
