% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_weighted_analysis.R
\name{do_weighted_cif_analysis}
\alias{do_weighted_cif_analysis}
\title{Run Weighted CIF Analysis}
\usage{
do_weighted_cif_analysis(prepped, tau)
}
\arguments{
\item{prepped}{A list returned by \code{prep_data_weighted_cif()}.}

\item{tau}{Numeric time horizon (e.g., 15 or 29).}
}
\value{
A list with \code{$WRMLT} and \code{$WRMGT}, each a 3x4 matrix.
}
\description{
Given the list from \code{\link{prep_data_weighted_cif}}, which contains
  \code{Treatment.death, Control.death, Treatment.discharge, Control.discharge}, run Weighted RMLT (death)
  and Weighted RMGT (discharge) at a user-specified time horizon \code{tau}.
}
\details{
- Weighted RMLT uses \code{eta=2}.
- Weighted RMGT uses \code{eta=1}.
}
