% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cif.R
\name{I}
\alias{I}
\title{Indicator function}
\usage{
I(x)
}
\arguments{
\item{x}{A logical value.}
}
\description{
A small helper that returns 1 if expression is TRUE, otherwise 0.
Used internally for clarity.
}
