% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_cif.R
\name{prep_data_cif}
\alias{prep_data_cif}
\title{Prepare Data (Part I: Standard CIF Only)}
\usage{
prep_data_cif(
  path_ch = NULL,
  col_map = list(ttrecov = "TTRECOV", ttdeath = "TTDEATH", reccnsr = "RECCNSR", dthcnsr =
    "DTHCNSR", ordscr_bs = "ordscr_bs", trt = "trt")
)
}
\arguments{
\item{path_ch}{(Optional) Path to the wide-format CSV. If \code{NULL},
the built-in dummy data (\code{dummy_actt_ch}) is used.}

\item{col_map}{Named list mapping columns in the wide dataset to needed variables:
\describe{
  \item{\code{ttrecov}}{Time to recovery.}
  \item{\code{ttdeath}}{Time to death.}
  \item{\code{reccnsr}}{Indicator for censored/not for recovery.}
  \item{\code{dthcnsr}}{Indicator for censored/not for death.}
  \item{\code{ordscr_bs}}{Baseline ordinal score.}
  \item{\code{trt}}{Treatment group. (0=control, 1=treatment)}
}}
}
\value{
A list with:
\itemize{
  \item \code{df_ch}: the cleaned wide dataframe.
  \item \code{data.w}: the final dataset for standard CIF with columns
        (\code{USUBJID, etime, estatus, etype2, trt}).
  \item \code{trt1}: subset of \code{data.w} where \code{trt==1}.
  \item \code{trt0}: subset of \code{data.w} where \code{trt==0}.
}
}
\description{
This function reads in (or uses built-in) wide-format data and
prepares it for standard CIF analysis (Part I).
}
