% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_wrt_cif.R
\name{prep_data_wrt_cif}
\alias{prep_data_wrt_cif}
\title{Prepare Data (Part II: Weighted CIF)}
\usage{
prep_data_wrt_cif(
  part1_output,
  path_long = NULL,
  col_map_long = list(adyc = "ADYC", ordscor = "ORDSCOR")
)
}
\arguments{
\item{part1_output}{The list returned by \code{prep_data_part1()}, containing
\code{df_ch} among other elements.}

\item{path_long}{(Optional) Path to the long-format CSV. If \code{NULL}, uses
\code{dummy_actt_long} instead (for demonstration).}

\item{col_map_long}{A named list specifying columns in the long data:
\describe{
  \item{\code{adyc}}{Measurement day/time.}
  \item{\code{ordscor}}{Ordinal score.}
}}
}
\value{
A list with the Weighted CIF elements:
\itemize{
  \item \code{data.ws.death}, \code{data.ws.discharge}: full wide+long merges.
  \item \code{trt1.death}, \code{trt0.death}: for event=death subsets.
  \item \code{trt1.discharge}, \code{trt0.discharge}: for event=discharge subsets.
}
}
\description{
Extends the cleaned wide data (from \code{\link{prep_data_part1}}) by merging
a long-format dataset and computing weighted times for death and discharge.
This is only relevant if the user has repeated measurements (e.g., ordinal scores).
}
