% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cif.R
\name{var_est}
\alias{var_est}
\title{Influence Function for Weighted Survival}
\usage{
var_est(data, t, eta)
}
\arguments{
\item{data}{Same structure as \code{\link{w_surv}}.}

\item{t}{Cutoff for \code{wU}.}

\item{eta}{Event type of interest.}
}
\value{
A numeric vector of length = number of rows in \code{data}, containing the influence function for each subject.
}
\description{
Calculates the influence function (a_j) for each individual, used in variance estimation.
}
