% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cif.R
\name{w_surv}
\alias{w_surv}
\title{Weighted survival function for competing risks}
\usage{
w_surv(data, t, eta, m)
}
\arguments{
\item{data}{A data frame that must include:
\describe{
  \item{D_time}{Time to event}
  \item{D_status}{Censoring indicator (1 = event, 0 = censored)}
  \item{etype}{Event type}
  \item{wU}{Weighted survival time}
}}

\item{t}{A numeric value, the cutoff for \code{wU}.}

\item{eta}{The event type of interest (1 or 2).}

\item{m}{The upper bound on \code{D_time}.}
}
\value{
A single numeric value of the weighted survival function at the given cutoff.
}
\description{
Estimates \eqn{E[I(U \le t, \eta = j, X \le m)]} given a data set with
weighted survival times.
}
